requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
MLB-StatsAPI==0.0.8
PyJWT==1.7.1
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
requests==2.22.0
six==1.12.0
twilio==6.27.0
urllib3==1.25.3
